app/requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
protobuf==3.14.0
pydgraph
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
